Return to starting directory and builld .tar.gz if can't package app

+++ .scripts/build_packages.sh
@@ -21,7 +21,14 @@ do
     cd $pkg_dir
     python setup.py sdist
     resilient-sdk package -p .
-    mv dist/app-* $dist_dir
+    if [ -f dist/app* ]; then
+        mv dist/app-* $dist_dir
+    else
+        # if couldn't build app.zip - build tar.gz
+        mv dist/*.tar.gz $dist_dir
+    fi
+    # return to the starting directory
+    cd $(dirname $0)
 done;
 
 # Build content packages with resilient-res-package.sh
